+++ packages/hotspot/configure/src/__tests__/index.test.js
@@ -2,6 +2,17 @@ import { shallow } from 'enzyme';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
+jest.mock('@pie-lib/config-ui', () => ({
+  choiceUtils: {
+    firstAvailableIndex: jest.fn()
+  },
+  settings: {
+    Panel: props => <div {...props} />,
+    toggle: jest.fn(),
+    radio: jest.fn()
+  }
+}));
+
 const model = () => ({
   prompt: 'This is the question prompt',
   imageUrl: '',
@@ -65,22 +76,19 @@ describe('index', () => {
   });
 
   describe('logic', () => {
-    describe('onMultipleCorrectChanged', () => {
-      it('resets the model', () => {
-        el.onMultipleCorrectChanged();
-
-        expect(onModelChanged).toBeCalledWith(
-          expect.objectContaining({ multipleCorrect: false }),
-        );
-      });
-    });
-
-    describe('onPartialScoringChanged', () => {
+    describe('onModelChangedByConfig', () => {
       it('changes partial scoring value', () => {
-        el.onPartialScoringChanged();
+        el.onModelChangedByConfig({ ...initialModel, partialScoring: true });
 
         expect(onModelChanged).toBeCalledWith(
           expect.objectContaining({ partialScoring: true }),
+        );
+      });
+      it('changes multiple correct value', () => {
+        el.onModelChangedByConfig({ ...initialModel, multipleCorrect: false });
+
+        expect(onModelChanged).toBeCalledWith(
+          expect.objectContaining({ multipleCorrect: false }),
         );
       });
     });
@@ -114,6 +122,17 @@ describe('index', () => {
 
         expect(onModelChanged).toBeCalledWith(
           expect.objectContaining({ prompt: newPrompt }),
+        );
+      });
+    });
+
+    describe('onRationaleChanged', () => {
+      it('changes the rationale', () => {
+        const newRationale = 'New Rationale';
+        el.onRationaleChanged(newRationale);
+
+        expect(onModelChanged).toBeCalledWith(
+          expect.objectContaining({ rationale: newRationale }),
         );
       });
     });