+++ packages/match/configure/src/configure.jsx
@@ -3,8 +3,10 @@ import { withStyles } from '@material-ui/core/styles';
 import {
   FeedbackConfig,
   settings,
-  layout
+  layout,
+  InputContainer
 } from '@pie-lib/config-ui';
+import EditableHtml from '@pie-lib/editable-html';
 import PropTypes from 'prop-types';
 import debug from 'debug';
 import Typography from '@material-ui/core/Typography';
@@ -23,7 +25,15 @@ const styles = theme => ({
   },
   content: {
     marginTop: theme.spacing.unit * 2
-  }
+  },
+  promptHolder: {
+    width: '100%',
+    paddingTop: theme.spacing.unit * 2
+  },
+  prompt: {
+    paddingTop: theme.spacing.unit * 2,
+    width: '100%'
+  },
 });
 
 class Configure extends React.Component {
@@ -150,6 +160,20 @@ class Configure extends React.Component {
     this.props.onModelChanged(this.props.model);
   };
 
+  onPromptChanged = prompt => {
+    this.props.onModelChanged({
+      ...this.props.model,
+      prompt
+    });
+  };
+
+  onRationaleChanged = rationale => {
+    this.props.onModelChanged({
+      ...this.props.model,
+      rationale
+    });
+  };
+
   render() {
     const { classes, model, imageSupport, onModelChanged, configuration, onConfigurationChanged } = this.props;
     const {
@@ -157,9 +181,10 @@ class Configure extends React.Component {
       partialScoring,
       teacherInstructions,
       studentInstructions,
-      rationale,
+      rationale = {},
       lockChoiceOrder,
-      scoringType
+      scoringType,
+      prompt,
     } = configuration;
 
     log('[render] model', model);
@@ -203,6 +228,32 @@ class Configure extends React.Component {
               This interaction allows for either one or more correct answers. Setting more than one answer as correct allows for partial credit <i>(see the Scoring tab)</i>.
             </span>
           </Typography>
+          {prompt.settings && (
+            <InputContainer
+              label={prompt.label}
+              className={classes.promptHolder}
+            >
+              <EditableHtml
+                className={classes.prompt}
+                markup={model.prompt}
+                onChange={this.onPromptChanged}
+                imageSupport={imageSupport}
+                nonEmpty={!prompt.settings}
+                disableUnderline
+              />
+            </InputContainer>
+          )}
+          {rationale.enabled && (
+            <InputContainer label={rationale.label}
+                            className={classes.promptHolder}>
+              <EditableHtml
+                className={classes.prompt}
+                markup={model.rationale || ''}
+                onChange={this.onRationaleChanged}
+                imageSupport={imageSupport}
+              />
+            </InputContainer>
+          )}
           <GeneralConfigBlock
             model={model}
             configuration={configuration}