+++ packages/pie-models/src/pie/match/index.ts
@@ -1,6 +1,6 @@
-import {PromptConfig} from '../../PromptConfig';
-import {CommonConfigSettings} from '../../CommonConfigSettings';
-import {PieModel} from '../../PieModel';
+import { PromptConfig } from '../../PromptConfig';
+import { CommonConfigSettings } from '../../CommonConfigSettings';
+import { PieModel } from '../../PieModel';
 import { ComplexFeedbackType } from '../../Feedback';
 import { ConfigureProp } from '../ConfigurationProp';
 
@@ -16,21 +16,22 @@ export interface MatchRow {
   values: boolean[];
 }
 
-/** NOTE: teacherInstructions, studentInstructions, rationale & scoringType
+/**
+ * NOTE: teacherInstructions, studentInstructions, rationale & scoringType
  * functionalities are not defined yet - the value for those can belong to
  * model or to configure (to be moved when the functionality is defined)
  */
 
 /**
-* Model for the @pie-elements/match Interaction
-* @additionalProperties false
-*/
+ * Model for the @pie-elements/match Interaction
+ * @additionalProperties false
+ */
 export interface MatchPie extends PieModel {
   /** Indicates if questions can contain images */
   enableImages?: boolean;
 
   /** Feedback for student responses */
-  feedback?: ComplexFeedbackType
+  feedback?: ComplexFeedbackType;
 
   /** Array of strings for column headers */
   headers: string[];
@@ -39,28 +40,32 @@ export interface MatchPie extends PieModel {
   layout: number;
 
   /** Indicates if the order of the rows should be randomly sorted on render */
-  lockChoiceOrder: number;
+  lockChoiceOrder?: boolean;
 
   /** Indicates if partial scoring should be used */
   partialScoring?: boolean;
 
   /** Indicates if the control for responses should be single (radio) or multiple (checkbox) */
-  choiceMode: 'radio' | 'checkbox';
+  choiceMode?: 'radio' | 'checkbox';
+
+  /**  The question prompt or item stem */
+  prompt: string;
 
   /** The rows of choices to be presented.  */
   rows: MatchRow[];
 
-  /** Indicates if rationale is enabled */
-  rationale: boolean;
+  /** Indicates value for rationale */
+
+  rationale: string;
 
   /** Indicates scoring type */
-  scoringType: 'auto' | 'rubric';
+  scoringType?: 'auto' | 'rubric';
 
   /** Indicates if student instructions are enabled */
-  studentInstructions: boolean;
+  studentInstructions?: boolean;
 
   /** Indicates if teacher instructions are enabled */
-  teacherInstructions: boolean;
+  teacherInstructions?: boolean;
 }
 
 /**
@@ -71,38 +76,42 @@ export interface MatchConfigure extends PromptConfig, CommonConfigSettings {
   /**
    * Configuration for enable images
    */
-  enableImages: ConfigureProp;
+  enableImages?: ConfigureProp;
 
   /**
    * Configuration for feedback
    */
-  feedback: ConfigureProp;
+  feedback?: ConfigureProp;
 
   /**
    * Configuration for headers
    */
-  headers: ConfigureProp;
+  headers?: ConfigureProp;
 
   /**
    * Configuration for layout
    */
-  layout: ConfigureProp;
+  layout?: ConfigureProp;
 
   /**
    * Configuration for lock choice order
    */
-  lockChoiceOrder: ConfigureProp;
+  lockChoiceOrder?: ConfigureProp;
 
   /**
    * Configuration for partial scoring
    */
-  partialScoring: ConfigureProp;
+  partialScoring?: ConfigureProp;
 
   /**
    * Configuration for choice mode
    */
-  choiceMode: ConfigureProp;
+  choiceMode?: ConfigureProp;
 
+  /**
+   * Configuration for the prompt
+   */
+  prompt?: ConfigureProp;
 
   /**
    * Rationale configuration