+++ packages/placement-ordering/controller/src/index.js
@@ -126,6 +126,12 @@ export function model(question, session, env) {
 
     base.disabled = env.mode !== 'gather';
 
+    if (env.role === 'instructor' && (env.mode === 'view' || env.mode === 'evaluate')) {
+      base.rationale = question.rationale;
+    } else {
+      base.rationale = null;
+    }
+
     if (env.mode === 'evaluate') {
       base.outcomes = _.map(session.value, function(c, idx) {
         return {